+++ example/evm-census/index.ts
@@ -158,9 +158,9 @@ async function launchNewVote() {
 
     console.log("Getting the block height")
     const currentBlock = await VotingApi.getBlockHeight(pool)
-    const startBlock = currentBlock + 25
+    const startBlock = currentBlock + 15
     // const blockCount = 6 * 10 // 10m
-    const blockCount = 6 * 4 // 4m
+    const blockCount = 6 * 1 // 4m
 
     // TODO: COMPUTE THE PARAMS SIGNATURE
     // TODO: INCLUDE THE BALANCE SLOT IN SUCH SIGNATURE